Onboarding: unify hook context and share adapter test helper
